Refactor presenters.
Bind interactors in base presenter. Unsuscribe them all in onDestroy.
Move attach/detach input listeners to BaseInteractor.
